refactor: changed GrantList folder name to KarmaGrant
